+++ src/Plugin.php
@@ -18,10 +18,10 @@ use craft\commerce\exports\LineItemExport;
 use craft\commerce\exports\OrderExport;
 use craft\commerce\fields\Products;
 use craft\commerce\fields\Variants;
+use craft\commerce\gql\interfaces\elements\Product as GqlProductInterface;
 use craft\commerce\gql\interfaces\elements\Variant as GqlVariantInterface;
 use craft\commerce\gql\queries\Product as GqlProductQueries;
 use craft\commerce\gql\queries\Variant as GqlVariantQueries;
-use craft\commerce\gql\interfaces\elements\Product as GqlProductInterface;
 use craft\commerce\helpers\ProjectConfigData;
 use craft\commerce\migrations\Install;
 use craft\commerce\models\Settings;
@@ -156,7 +156,6 @@ class Plugin extends BasePlugin
         $this->_registerRedactorLinkOptions();
         $this->_registerPermissions();
         $this->_registerCraftEventListeners();
-        $this->_registerSessionEventListeners();
         $this->_registerProjectConfigEventListeners();
         $this->_registerWidgets();
         $this->_registerVariables();
@@ -172,7 +171,7 @@ class Plugin extends BasePlugin
         $this->_registerElementExports();
         $this->_defineResaveCommand();
 
-        Craft::setAlias('@commerceLib',  Craft::getAlias('@craft/commerce/../lib'));
+        Craft::setAlias('@commerceLib', Craft::getAlias('@craft/commerce/../lib'));
     }
 
     /**
@@ -379,18 +378,6 @@ class Plugin extends BasePlugin
     }
 
     /**
-     * Register Commerce’s session event listeners
-     */
-    private function _registerSessionEventListeners()
-    {
-        if (!Craft::$app->getRequest()->getIsConsoleRequest()) {
-            Event::on(UserElement::class, UserElement::EVENT_AFTER_SAVE, [$this->getCustomers(), 'saveUserHandler']);
-            Event::on(User::class, User::EVENT_AFTER_LOGIN, [$this->getCustomers(), 'loginHandler']);
-            Event::on(User::class, User::EVENT_AFTER_LOGOUT, [$this->getCustomers(), 'logoutHandler']);
-        }
-    }
-
-    /**
      * Register Commerce’s project config event listeners
      */
     private function _registerProjectConfigEventListeners()
@@ -447,9 +434,14 @@ class Plugin extends BasePlugin
      */
     private function _registerCraftEventListeners()
     {
+        if (!Craft::$app->getRequest()->isConsoleRequest) {
+            Event::on(User::class, User::EVENT_AFTER_LOGIN, [$this->getCustomers(), 'loginHandler']);
+            Event::on(User::class, User::EVENT_AFTER_LOGOUT, [$this->getCustomers(), 'logoutHandler']);
+        }
+
         Event::on(Sites::class, Sites::EVENT_AFTER_SAVE_SITE, [$this->getProductTypes(), 'afterSaveSiteHandler']);
         Event::on(Sites::class, Sites::EVENT_AFTER_SAVE_SITE, [$this->getProducts(), 'afterSaveSiteHandler']);
-        Event::on(UserElement::class, UserElement::EVENT_AFTER_SAVE, [$this->getCustomers(), 'afterSaveUserHandler']);
+        Event::on(UserElement::class, UserElement::EVENT_AFTER_SAVE, [$this->getCustomers(), 'afterSaveUserHandler'], null, false); // Lets run this before other plugins if we can
         Event::on(UserElement::class, UserElement::EVENT_BEFORE_DELETE, [$this->getSubscriptions(), 'beforeDeleteUserHandler']);
         Event::on(Purchasable::class, Elements::EVENT_BEFORE_RESTORE_ELEMENT, [$this->getPurchasables(), 'beforeRestorePurchasableHandler']);
     }