+++ src/migrations/m191115_103105_add_totalDiscountUses_column.php
@@ -7,7 +7,6 @@
 
 namespace craft\commerce\migrations;
 
-use Craft;
 use craft\db\Migration;
 
 /**