+++ src/models/ShippingMethodOption.php
@@ -7,20 +7,25 @@
 
 namespace craft\commerce\models;
 
-use Craft;
+use craft\commerce\behaviors\CurrencyAttributeBehavior;
 use craft\commerce\elements\Order;
+use craft\commerce\Plugin;
+use yii\behaviors\AttributeTypecastBehavior;
 
 /**
  * Shipping method option model.
  *
  * @property float $price the price of the shipping method option for the order.
+ * @property Order $order
+ * @property string $currency
+ * @property-read string $priceAsCurrency
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 3.1
  */
 class ShippingMethodOption extends ShippingMethod
 {
     /**
-     * @var mixed
+     * @var Order
      */
     private $_order;
 
@@ -32,20 +37,24 @@ class ShippingMethodOption extends ShippingMethod
     /**
      * @return array
      */
-    public function fields(): array
+    public function behaviors(): array
     {
-        $fields = parent::fields();
+        $behaviors = parent::behaviors();
 
-        foreach ($this->currencyAttributes() as $attribute) {
-            $fields[$attribute . 'AsCurrency'] = function($model, $attribute) {
-                // Substr because attribute is returned with 'AsCurrency' appended
-                $attribute = substr($attribute, 0, -10);
-                $amount = $model->$attribute ?? 0;
-                return Craft::$app->getFormatter()->asCurrency($amount, $this->_order->currency, [], [], true);
-            };
-        }
+        $behaviors['typecast'] = [
+            'class' => AttributeTypecastBehavior::class,
+            'attributeTypes' => [
+                'id' => AttributeTypecastBehavior::TYPE_INTEGER
+            ]
+        ];
 
-        return $fields;
+        $behaviors['currencyAttributes'] = [
+            'class' => CurrencyAttributeBehavior::class,
+            'defaultCurrency' => Plugin::getInstance()->getPaymentCurrencies()->getPrimaryPaymentCurrencyIso(),
+            'currencyAttributes' => $this->currencyAttributes()
+        ];
+
+        return $behaviors;
     }
 
     /**
@@ -57,8 +66,15 @@ class ShippingMethodOption extends ShippingMethod
     {
         $attributes = [];
         $attributes[] = 'price';
-
         return $attributes;
+    }
+
+    /**
+     * @return string
+     */
+    protected function getCurrency(): string
+    {
+        return $this->_order->currency ?? parent::getCurrency();
     }
 
     /**
@@ -71,7 +87,7 @@ class ShippingMethodOption extends ShippingMethod
 
     /**
      * @param $order
-     * @since 3.x
+     * @since 3.1.10
      */
     public function setOrder($order)
     {