+++ src/services/Customers.php
@@ -23,7 +23,6 @@ use craft\elements\User;
 use craft\elements\User as UserElement;
 use craft\errors\ElementNotFoundException;
 use craft\events\ModelEvent;
-use craft\events\UserEvent as CraftUserEvent;
 use Throwable;
 use Twig\Error\LoaderError;
 use Twig\Error\RuntimeError;
@@ -47,7 +46,6 @@ use yii\web\UserEvent;
 class Customers extends Component
 {
     const SESSION_CUSTOMER = 'commerce_customer';
-
 
     /**
      * @var Customer
@@ -134,6 +132,16 @@ class Customers extends Component
             $this->_customer = $customer;
         }
 
+        // Always ensure the customer is saved.
+        if (!$this->_customer->id) {
+            if ($this->saveCustomer($this->_customer)) {
+                Craft::$app->getSession()->set(self::SESSION_CUSTOMER, $this->_customer->id);
+            } else {
+                $errors = implode(', ', $this->_customer->errors);
+                throw new Exception('Error saving customer: ' . $errors);
+            }
+        }
+
         return $this->_customer;
     }
 
@@ -150,7 +158,7 @@ class Customers extends Component
     {
         // default to customer in session.
         if (null === $customer) {
-            $customer = $this->_getSavedCustomer();
+            $customer = $this->getCustomer();
         }
 
         if (Plugin::getInstance()->getAddresses()->saveAddress($address, $runValidation)) {
@@ -346,7 +354,7 @@ class Customers extends Component
                 // Only consolidate completed orders, not carts and orders that don't belong to another user.
 
                 if ($order->isCompleted && !$belongsToAnotherUser) {
-                    $order->customerId = $toCustomer->id;
+                    $order->setCustomer($toCustomer);
 
                     // We only want to update search indexes if the order is a cart and the developer wants to keep cart search indexes updated.
                     $updateCartSearchIndexes = Plugin::getInstance()->getSettings()->updateCartSearchIndexes;
@@ -440,10 +448,13 @@ class Customers extends Component
      *
      * @return int
      * @throws Exception
+     * @deprecated 3.x
      */
     public function getCustomerId(): int
     {
-        return $this->_getSavedCustomer()->id;
+        Craft::$app->getDeprecator()->log('Customers::getCustomerId()', 'Customers::getCustomerId() has been deprecated. Use Customers::getCustomer()->id, since it is guaranteed to have a ID.');
+
+        return $this->getCustomer()->id;
     }
 
     /**
@@ -451,24 +462,14 @@ class Customers extends Component
      *
      * @param Event $event
      * @throws Exception
+     * @deprecated 3.x
      */
     public function saveUserHandler(Event $event)
     {
-        $user = $event->sender;
-        $customer = $this->getCustomerByUserId($user->id);
-        $email = $user->email;
+        Craft::$app->getDeprecator()->log('Customers::saveUserHandler()', 'Customers::saveUserHandler() has been deprecated. Use Customers::afterSaveUserHandler() instead.');
 
-        // Update the email address on orders for this customer.
-        if ($customer) {
-            $orders = (new Query())
-                ->select(['orders.id'])
-                ->from([Table::ORDERS . ' orders'])
-                ->where(['orders.customerId' => $customer->id])
-                ->column();
-
-            Craft::$app->getDb()->createCommand()
-                ->update(Table::ORDERS, ['email' => $email], ['id' => $orders])
-                ->execute();
+        if ($customer = $this->getCustomerByUserId($event->sender->id)) {
+            $this->_updatePreviousOrderEmails($customer->id, $event->sender->email);
         }
     }
 
@@ -498,7 +499,7 @@ class Customers extends Component
                 'primaryShippingAddressId',
             ])
             ->from(Table::CUSTOMERS . ' customers')
-            ->innerJoin(Table::ORDERS . ' orders' , '[[orders.customerId]] = [[customers.id]]')
+            ->innerJoin(Table::ORDERS . ' orders', '[[orders.customerId]] = [[customers.id]]')
             ->leftJoin(CraftTable::USERS . ' users', '[[users.id]] = [[customers.userId]]')
             ->leftJoin(Table::ADDRESSES . ' billing', '[[billing.id]] = [[customers.primaryBillingAddressId]]')
             ->leftJoin(Table::ADDRESSES . ' shipping', '[[shipping.id]] = [[customers.primaryShippingAddressId]]')
@@ -516,16 +517,27 @@ class Customers extends Component
             ])
 
             // Exclude customer records without a user or where there isn't any data
-            ->where(['or',
+            ->where([
+                'or',
                 ['not', ['userId' => null]],
-                ['and',
+                [
+                    'and',
                     ['userId' => null],
-                    ['or',
+                    [
+                        'or',
                         ['not', ['primaryBillingAddressId' => null]],
                         ['not', ['primaryShippingAddressId' => null]],
                     ]
                 ]
-            ])->andWhere(['[[orders.isCompleted]]' => 1]);
+            ])->andWhere([
+                'or',
+                ['orders.isCompleted' => true],
+                [
+                    'and',
+                    ['orders.isCompleted' => false],
+                    ['not', ['customers.userId' => null]],
+                ]
+            ]);
 
         if ($search) {
             $likeOperator = Craft::$app->getDb()->getIsPgsql() ? 'ILIKE' : 'LIKE';
@@ -591,7 +603,11 @@ class Customers extends Component
             if (!$userId) {
                 // Dont use element save, just update DB directly
                 if ($order && $order instanceof Order) {
-                    $order->customerId = $customerId;
+                    $customer = Plugin::getInstance()->getCustomers()->getCustomerById($customerId);
+                    if (!$customer) {
+                        throw new Exception('Can’t find customer to assign order to');
+                    }
+                    $order->setCustomer($customer);
                 }
 
                 Craft::$app->getDb()->createCommand()
@@ -603,28 +619,6 @@ class Customers extends Component
                     ->execute();
             }
         }
-    }
-
-    /**
-     * Get the current customer.
-     *
-     * @return Customer
-     * @throws Exception if failed to save customer.
-     */
-    private function _getSavedCustomer(): Customer
-    {
-        $customer = $this->getCustomer();
-
-        if (!$customer->id) {
-            if ($this->saveCustomer($customer)) {
-                Craft::$app->getSession()->set(self::SESSION_CUSTOMER, $customer->id);
-            } else {
-                $errors = implode(', ', $customer->errors);
-                throw new Exception('Error saving customer: ' . $errors);
-            }
-        }
-
-        return $customer;
     }
 
     /**
@@ -869,27 +863,57 @@ class Customers extends Component
      */
     public function afterSaveUserHandler(ModelEvent $event)
     {
-        // Check to see if there is a customer record for the user
-        $customer = $this->getCustomerByUserId($event->sender->id);
+        $user = $event->sender;
+        $customer = $this->getCustomerByUserId($user->id);
+        $email = $user->email;
 
         // Create a new customer for a user that does not have a customer
-        if (!$customer && $event->sender->email) {
+        if (!$customer && $user->email) {
             $existingCustomerIdByEmail = (new Query())
                 ->select('orders.customerId')
                 ->from(Table::ORDERS . ' orders')
                 ->innerJoin(Table::CUSTOMERS . ' customers', '[[customers.id]] = [[orders.customerId]]')
-                ->where(['orders.email' => $event->sender->email])
+                ->where(['orders.email' => $user->email])
                 ->andWhere(['orders.isCompleted' => true])
                 ->orderBy('[[orders.dateOrdered]] ASC')
                 ->scalar(); // get the first customerId in the result
 
             if ($customer = $this->getCustomerById($existingCustomerIdByEmail)) {
-                $customer->userId = $event->sender->id;
+                $customer->userId = $user->id;
             } else {
-                $customer = new Customer(['userId' => $event->sender->id]);
+                $customer = new Customer(['userId' => $user->id]);
             }
 
             $this->saveCustomer($customer);
         }
+
+        // Update the email address in the DB for this customer
+        if ($email) {
+            $this->_updatePreviousOrderEmails($customer->id, $email);
+        }
+
+        $this->consolidateGuestOrdersByEmail($email);
     }
+
+    /**
+     * @param $customerId
+     * @param $email
+     * @throws \yii\db\Exception
+     */
+    private function _updatePreviousOrderEmails(int $customerId, string $email)
+    {
+        $orderIds = (new Query())
+            ->select(['orders.id'])
+            ->from([Table::ORDERS . ' orders'])
+            ->where(['orders.customerId' => $customerId])
+            ->andWhere(['not', ['orders.email' => $email]])
+            ->column();
+
+        if (!empty($orderIds)) {
+            Craft::$app->getDb()->createCommand()
+                ->update(Table::ORDERS, ['email' => $email], ['id' => $orderIds])
+                ->execute();
+        }
+    }
+
 }