+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Release Notes for Craft Commerce
 
+## 3.1.11 - 2020-07-06
+
+### Added
+- Added new `*AsCurrency` attributes to all currency attributes on orders, line items, products, variants, adjustments and transactions.
+- Added the `hasVariant` argument to GraphQL product queries. ([#1544](https://github.com/craftcms/commerce/issues/1544))
+- Added `craft\commerce\events\ModifyCartInfoEvent::$cart`. ([#1536](https://github.com/craftcms/commerce/issues/1536))
+- Added `craft\commerce\behaviors\CurrencyAttributeBehavior`.
+- Added `craft\commerce\gql\types\input\Variant`.
+
+### Fixed
+- Improved performance when adding items to the cart. ([#1543](https://github.com/craftcms/commerce/pull/1543), [#1520](https://github.com/craftcms/commerce/issues/1520))
+- Fixed a bug where products that didn’t have current sales could be returned when the `hasSales` query parameter was enabled.
+- Fixed a bug where the “Message” field wasn’t getting cleared after updating the order status on the Order edit page. ([#1366](https://github.com/craftcms/commerce/issues/1366))
+- Fixed a bug where it wasn’t possible to update the conversion rate on a payment currency. ([#1547](https://github.com/craftcms/commerce/issues/1547))
+- Fixed a bug where it wasn’t possible to delete all line item statuses.
+- Fixed a bug where zero currency values weren’t getting formatted correctly in `commerce/cart/*` actions’ JSON responses. ([#1539](https://github.com/craftcms/commerce/issues/1539))
+- Fixed a bug where the wrong line item could be added to the cart when using the Lite edition. ([#1552](https://github.com/craftcms/commerce/issues/1552))
+- Fixed a bug where a validation error was being shown incorrectly on the Edit Discount page. ([#1549](https://github.com/craftcms/commerce/issues/1549))
+
+### Deprecated
+- The `|json_encode_filtered` twig filter has now been deprecated.
+
 ## 3.1.10 - 2020-06-23
 
 ### Added